refactor(service-food-crawler): minor changes
