--- service-food-crawler/src/main/kotlin/de/hpi/cloud/food/OpenMensaInformation.kt
@@ -1,9 +1,0 @@
-package de.hpi.cloud.food
-
-data class OpenMensaInformation(
-    val id: Int,
-    val name: String,
-    val city: String,
-    val address: String,
-    val coordinates: List<Double>
-) {}
