--- service-food-crawler/src/main/kotlin/de/hpi/cloud/food/StaticCanteenData.kt
@@ -1,7 +1,0 @@
-package de.hpi.cloud.food
-
-data class StaticCanteenData(
-    val canteenOpenMensaId: Int,
-    val canteenId: String,
-    val counterFinder: (OpenMensaMeal) -> String?
-) {}
