+++ service-food-crawler/src/main/kotlin/de/hpi/cloud/food/crawler/utils/UrlUtils.kt
@@ -1,0 +1,11 @@
+package de.hpi.cloud.food.crawler.utils
+
+import java.net.URL
+
+fun URL.openConnectionWith(vararg properties: Pair<String, String>) =
+    openConnection()!!
+        .apply {
+            properties.forEach { setRequestProperty(it.first, it.second) }
+        }
+
+fun URL.openStreamWith(vararg properties: Pair<String, String>) = openConnectionWith(*properties).inputStream!!
